tests for adapter fail
